change code

+++ gdt.c
@@ -1,6 +1,6 @@
 #include "kernel.h"
 
-uint64_t create_descriptor(uint32_t base, uint32_t limit, uint16_t flag)
+uint64_t create_descriptor(uint64_t base, uint64_t limit, uint16_t flag)
 {
     uint64_t descriptor;
